Moved Epoch Python to its own file
