+++ src/epoch/mod.rs
@@ -30,16 +30,10 @@ use core::hash::{Hash, Hasher};
 use core::ops::{Add, AddAssign, Sub, SubAssign};
 
 #[cfg(feature = "python")]
+mod python;
+
+#[cfg(feature = "python")]
 use pyo3::prelude::*;
-
-#[cfg(feature = "python")]
-use pyo3::pyclass::CompareOp;
-
-#[cfg(feature = "python")]
-use pyo3::types::PyType;
-
-#[cfg(feature = "python")]
-use crate::leap_seconds_file::LeapSecondsFile;
 
 #[cfg(not(kani))]
 #[cfg(feature = "serde")]
@@ -48,7 +42,7 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use core::str::FromStr;
 
 #[cfg(feature = "ut1")]
-use crate::ut1::Ut1Provider;
+pub mod ut1;
 
 #[cfg(not(feature = "std"))]
 #[allow(unused_imports)] // Import is indeed used.
@@ -1233,23 +1227,6 @@ impl Epoch {
             .parse(s_in)
     }
 
-    #[cfg(feature = "ut1")]
-    #[must_use]
-    /// Initialize an Epoch from the provided UT1 duration since 1900 January 01 at midnight
-    ///
-    /// # Warning
-    /// The time scale of this Epoch will be set to TAI! This is to ensure that no additional computations will change the duration since it's stored in TAI.
-    /// However, this also means that calling `to_duration()` on this Epoch will return the TAI duration and not the UT1 duration!
-    pub fn from_ut1_duration(duration: Duration, provider: Ut1Provider) -> Self {
-        let mut e = Self::from_tai_duration(duration);
-        // Compute the TAI to UT1 offset at this time.
-        // We have the time in TAI. But we were given UT1.
-        // The offset is provided as offset = TAI - UT1 <=> TAI = UT1 + offset
-        e.duration += e.ut1_offset(provider).unwrap_or(Duration::ZERO);
-        e.time_scale = TimeScale::TAI;
-        e
-    }
-
     fn delta_et_tai(seconds: f64) -> f64 {
         // Calculate M, the mean anomaly.4
         let m = NAIF_M0 + seconds * NAIF_M1;
@@ -1448,502 +1425,6 @@ impl Epoch {
     /// This is to match the `iauDat` function of SOFA (src/dat.c). That function will return a warning and give up if the start date is before 1960.
     pub fn leap_seconds(&self, iers_only: bool) -> Option<f64> {
         self.leap_seconds_with(iers_only, LatestLeapSeconds::default())
-    }
-
-    #[cfg(feature = "ut1")]
-    /// Get the accumulated offset between this epoch and UT1, assuming that the provider includes all data.
-    pub fn ut1_offset(&self, provider: Ut1Provider) -> Option<Duration> {
-        for delta_tai_ut1 in provider.rev() {
-            if self > &delta_tai_ut1.epoch {
-                return Some(delta_tai_ut1.delta_tai_minus_ut1);
-            }
-        }
-        None
-    }
-
-    /// Get the accumulated number of leap seconds up to this Epoch from the provided LeapSecondProvider.
-    /// Returns None if the epoch is before 1960, year at which UTC was defined.
-    ///
-    /// # Why does this function return an `Option` when the other returns a value
-    /// This is to match the `iauDat` function of SOFA (src/dat.c). That function will return a warning and give up if the start date is before 1960.
-    #[cfg(feature = "python")]
-    pub fn leap_seconds_with_file(
-        &self,
-        iers_only: bool,
-        provider: LeapSecondsFile,
-    ) -> Option<f64> {
-        self.leap_seconds_with(iers_only, provider)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Creates a new Epoch from a Duration as the time difference between this epoch and TAI reference epoch.
-    const fn init_from_tai_duration(_cls: &Bound<'_, PyType>, duration: Duration) -> Self {
-        Self::from_tai_duration(duration)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Creates a new Epoch from its centuries and nanosecond since the TAI reference epoch.
-    fn init_from_tai_parts(_cls: &Bound<'_, PyType>, centuries: i16, nanoseconds: u64) -> Self {
-        Self::from_tai_parts(centuries, nanoseconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided TAI seconds since 1900 January 01 at midnight
-    fn init_from_tai_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_tai_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided TAI days since 1900 January 01 at midnight
-    fn init_from_tai_days(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_tai_days(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided UTC seconds since 1900 January 01 at midnight
-    fn init_from_utc_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_utc_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided UTC days since 1900 January 01 at midnight
-    fn init_from_utc_days(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_utc_days(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from given MJD in TAI time scale
-    fn init_from_mjd_tai(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_mjd_tai(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from given MJD in UTC time scale
-    fn init_from_mjd_utc(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_mjd_utc(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from given JDE in TAI time scale
-    fn init_from_jde_tai(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_jde_tai(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from given JDE in UTC time scale
-    fn init_from_jde_utc(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_jde_utc(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided TT seconds (approximated to 32.184s delta from TAI)
-    fn init_from_tt_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_tt_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided TT seconds (approximated to 32.184s delta from TAI)
-    fn init_from_tt_duration(_cls: &Bound<'_, PyType>, duration: Duration) -> Self {
-        Self::from_tt_duration(duration)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the Ephemeris Time seconds past 2000 JAN 01 (J2000 reference)
-    fn init_from_et_seconds(_cls: &Bound<'_, PyType>, seconds_since_j2000: f64) -> Epoch {
-        Self::from_et_seconds(seconds_since_j2000)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the Ephemeris Time duration past 2000 JAN 01 (J2000 reference)
-    fn init_from_et_duration(_cls: &Bound<'_, PyType>, duration_since_j2000: Duration) -> Self {
-        Self::from_et_duration(duration_since_j2000)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from Dynamic Barycentric Time (TDB) seconds past 2000 JAN 01 midnight (difference than SPICE)
-    /// NOTE: This uses the ESA algorithm, which is a notch more complicated than the SPICE algorithm, but more precise.
-    /// In fact, SPICE algorithm is precise +/- 30 microseconds for a century whereas ESA algorithm should be exactly correct.
-    fn init_from_tdb_seconds(_cls: &Bound<'_, PyType>, seconds_j2000: f64) -> Epoch {
-        Self::from_tdb_seconds(seconds_j2000)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from Dynamic Barycentric Time (TDB) (same as SPICE ephemeris time) whose epoch is 2000 JAN 01 noon TAI.
-    fn init_from_tdb_duration(_cls: &Bound<'_, PyType>, duration_since_j2000: Duration) -> Epoch {
-        Self::from_tdb_duration(duration_since_j2000)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from the JDE days
-    fn init_from_jde_et(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_jde_et(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from Dynamic Barycentric Time (TDB) (same as SPICE ephemeris time) in JD days
-    fn init_from_jde_tdb(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_jde_tdb(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of seconds since the GPS Time Epoch,
-    /// defined as UTC midnight of January 5th to 6th 1980 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29>).
-    fn init_from_gpst_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_gpst_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of days since the GPS Time Epoch,
-    /// defined as UTC midnight of January 5th to 6th 1980 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29>).
-    fn init_from_gpst_days(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_gpst_days(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of nanoseconds since the GPS Time Epoch,
-    /// defined as UTC midnight of January 5th to 6th 1980 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29>).
-    /// This may be useful for time keeping devices that use GPS as a time source.
-    fn init_from_gpst_nanoseconds(_cls: &Bound<'_, PyType>, nanoseconds: u64) -> Self {
-        Self::from_gpst_nanoseconds(nanoseconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of seconds since the QZSS Time Epoch,
-    /// defined as UTC midnight of January 5th to 6th 1980 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29>).
-    fn init_from_qzsst_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_qzsst_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of days since the QZSS Time Epoch,
-    /// defined as UTC midnight of January 5th to 6th 1980 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29>).
-    fn init_from_qzsst_days(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_qzsst_days(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of nanoseconds since the QZSS Time Epoch,
-    /// defined as UTC midnight of January 5th to 6th 1980 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS#GPS_Time_.28GPST.29>).
-    /// This may be useful for time keeping devices that use QZSS as a time source.
-    fn init_from_qzsst_nanoseconds(_cls: &Bound<'_, PyType>, nanoseconds: u64) -> Self {
-        Self::from_qzsst_nanoseconds(nanoseconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of seconds since the Galileo Time Epoch,
-    /// starting on August 21st 1999 Midnight UT,
-    /// (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS>).
-    fn init_from_gst_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_gst_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of days since the Galileo Time Epoch,
-    /// starting on August 21st 1999 Midnight UT,
-    /// (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS>).
-    fn init_from_gst_days(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_gst_days(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of nanoseconds since the Galileo Time Epoch,
-    /// starting on August 21st 1999 Midnight UT,
-    /// (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS>).
-    /// This may be useful for time keeping devices that use GST as a time source.
-    fn init_from_gst_nanoseconds(_cls: &Bound<'_, PyType>, nanoseconds: u64) -> Self {
-        Self::from_gst_nanoseconds(nanoseconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of seconds since the BeiDou Time Epoch,
-    /// defined as January 1st 2006 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS>).
-    fn init_from_bdt_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_bdt_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of days since the BeiDou Time Epoch,
-    /// defined as January 1st 2006 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS>).
-    fn init_from_bdt_days(_cls: &Bound<'_, PyType>, days: f64) -> Self {
-        Self::from_bdt_days(days)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the number of days since the BeiDou Time Epoch,
-    /// defined as January 1st 2006 (cf. <https://gssc.esa.int/navipedia/index.php/Time_References_in_GNSS>).
-    /// This may be useful for time keeping devices that use BDT as a time source.
-    fn init_from_bdt_nanoseconds(_cls: &Bound<'_, PyType>, nanoseconds: u64) -> Self {
-        Self::from_bdt_nanoseconds(nanoseconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided UNIX second timestamp since UTC midnight 1970 January 01.
-    fn init_from_unix_seconds(_cls: &Bound<'_, PyType>, seconds: f64) -> Self {
-        Self::from_unix_seconds(seconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize an Epoch from the provided UNIX millisecond timestamp since UTC midnight 1970 January 01.
-    fn init_from_unix_milliseconds(_cls: &Bound<'_, PyType>, milliseconds: f64) -> Self {
-        Self::from_unix_milliseconds(milliseconds)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    fn init_from_gregorian(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-        time_scale: TimeScale,
-    ) -> Self {
-        Self::from_gregorian(year, month, day, hour, minute, second, nanos, time_scale)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    fn init_from_gregorian_at_noon(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        time_scale: TimeScale,
-    ) -> Self {
-        Self::from_gregorian_at_noon(year, month, day, time_scale)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    fn init_from_gregorian_at_midnight(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        time_scale: TimeScale,
-    ) -> Self {
-        Self::from_gregorian_at_midnight(year, month, day, time_scale)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Attempts to build an Epoch from the provided Gregorian date and time in TAI.
-    fn maybe_init_from_gregorian_tai(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Result<Self, Errors> {
-        Self::maybe_from_gregorian_tai(year, month, day, hour, minute, second, nanos)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Attempts to build an Epoch from the provided Gregorian date and time in the provided time scale.
-    /// NOTE: If the time scale is TDB, this function assumes that the SPICE format is used
-    #[allow(clippy::too_many_arguments)]
-    fn maybe_init_from_gregorian(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-        time_scale: TimeScale,
-    ) -> Result<Self, Errors> {
-        Self::maybe_from_gregorian(year, month, day, hour, minute, second, nanos, time_scale)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Builds an Epoch from the provided Gregorian date and time in TAI. If invalid date is provided, this function will panic.
-    /// Use maybe_from_gregorian_tai if unsure.
-    fn init_from_gregorian_tai(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Self {
-        Self::from_gregorian_tai(year, month, day, hour, minute, second, nanos)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from the Gregorian date at midnight in TAI.
-    fn init_from_gregorian_tai_at_midnight(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-    ) -> Self {
-        Self::from_gregorian_tai_at_midnight(year, month, day)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from the Gregorian date at noon in TAI
-    fn init_from_gregorian_tai_at_noon(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-    ) -> Self {
-        Self::from_gregorian_tai_at_noon(year, month, day)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from the Gregorian date and time (without the nanoseconds) in TAI
-    fn init_from_gregorian_tai_hms(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-    ) -> Self {
-        Self::from_gregorian_tai_hms(year, month, day, hour, minute, second)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Attempts to build an Epoch from the provided Gregorian date and time in UTC.
-    fn maybe_init_from_gregorian_utc(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Result<Self, Errors> {
-        Self::maybe_from_gregorian_utc(year, month, day, hour, minute, second, nanos)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Builds an Epoch from the provided Gregorian date and time in TAI. If invalid date is provided, this function will panic.
-    /// Use maybe_from_gregorian_tai if unsure.
-    fn init_from_gregorian_utc(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-        nanos: u32,
-    ) -> Self {
-        Self::from_gregorian_utc(year, month, day, hour, minute, second, nanos)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from Gregorian date in UTC at midnight
-    fn init_from_gregorian_utc_at_midnight(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-    ) -> Self {
-        Self::from_gregorian_utc_at_midnight(year, month, day)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from Gregorian date in UTC at noon
-    fn init_from_gregorian_utc_at_noon(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-    ) -> Self {
-        Self::from_gregorian_utc_at_noon(year, month, day)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Initialize from the Gregorian date and time (without the nanoseconds) in UTC
-    fn init_from_gregorian_utc_hms(
-        _cls: &Bound<'_, PyType>,
-        year: i32,
-        month: u8,
-        day: u8,
-        hour: u8,
-        minute: u8,
-        second: u8,
-    ) -> Self {
-        Self::from_gregorian_utc_hms(year, month, day, hour, minute, second)
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    /// Equivalent to `datetime.strptime`, refer to <https://docs.rs/hifitime/latest/hifitime/efmt/format/struct.Format.html> for format options
-    fn strptime(_cls: &Bound<'_, PyType>, epoch_str: String, format_str: String) -> PyResult<Self> {
-        Self::from_format_str(&epoch_str, &format_str).map_err(|e| PyErr::from(e))
-    }
-
-    #[cfg(feature = "python")]
-    /// Equivalent to `datetime.strftime`, refer to <https://docs.rs/hifitime/latest/hifitime/efmt/format/struct.Format.html> for format options
-    fn strftime(&self, format_str: String) -> PyResult<String> {
-        use crate::efmt::Formatter;
-        let fmt = Format::from_str(&format_str)
-            .map_err(Errors::ParseError)
-            .map_err(|e| PyErr::from(e))?;
-        Ok(format!("{}", Formatter::new(*self, fmt)))
-    }
-
-    #[cfg(feature = "python")]
-    /// Equivalent to `datetime.isoformat`, and truncated to 23 chars, refer to <https://docs.rs/hifitime/latest/hifitime/efmt/format/struct.Format.html> for format options
-    fn isoformat(&self) -> PyResult<String> {
-        Ok(self.to_isoformat())
     }
 
     #[cfg(feature = "std")]
@@ -2475,21 +1956,6 @@ impl Epoch {
         Self::compute_gregorian(self.to_duration_in_time_scale(ts), ts)
     }
 
-    #[cfg(feature = "ut1")]
-    #[must_use]
-    /// Returns this time in a Duration past J1900 counted in UT1
-    pub fn to_ut1_duration(&self, provider: Ut1Provider) -> Duration {
-        // TAI = UT1 + offset <=> UTC = TAI - offset
-        self.to_tai_duration() - self.ut1_offset(provider).unwrap_or(Duration::ZERO)
-    }
-
-    #[cfg(feature = "ut1")]
-    #[must_use]
-    /// Returns this time in a Duration past J1900 counted in UT1
-    pub fn to_ut1(&self, provider: Ut1Provider) -> Self {
-        Self::from_tai_duration(self.to_ut1_duration(provider))
-    }
-
     #[must_use]
     /// Floors this epoch to the closest provided duration
     ///
@@ -2861,65 +2327,6 @@ impl Epoch {
     pub fn month_name(&self) -> MonthName {
         let month = Self::compute_gregorian(self.duration, self.time_scale).1;
         month.into()
-    }
-
-    // Python helpers
-
-    #[cfg(feature = "python")]
-    #[new]
-    fn new_py(string_repr: String) -> PyResult<Self> {
-        match Self::from_str(&string_repr) {
-            Ok(d) => Ok(d),
-            Err(e) => Err(PyErr::from(e)),
-        }
-    }
-
-    #[cfg(feature = "python")]
-    fn __getnewargs__(&self) -> Result<(String,), PyErr> {
-        Ok((format!("{self:?}"),))
-    }
-
-    #[cfg(feature = "python")]
-    #[classmethod]
-    fn system_now(_cls: &Bound<'_, PyType>) -> Result<Self, Errors> {
-        Self::now()
-    }
-
-    #[cfg(feature = "python")]
-    fn __str__(&self) -> String {
-        format!("{self}")
-    }
-
-    #[cfg(feature = "python")]
-    fn __repr__(&self) -> String {
-        format!("{self:?} @ {self:p}")
-    }
-
-    #[cfg(feature = "python")]
-    fn __add__(&self, duration: Duration) -> Self {
-        *self + duration
-    }
-
-    #[cfg(feature = "python")]
-    fn __sub__(&self, duration: Duration) -> Self {
-        *self - duration
-    }
-
-    #[cfg(feature = "python")]
-    fn timedelta(&self, other: Self) -> Duration {
-        *self - other
-    }
-
-    #[cfg(feature = "python")]
-    fn __richcmp__(&self, other: Self, op: CompareOp) -> bool {
-        match op {
-            CompareOp::Lt => *self < other,
-            CompareOp::Le => *self <= other,
-            CompareOp::Eq => *self == other,
-            CompareOp::Ne => *self != other,
-            CompareOp::Gt => *self > other,
-            CompareOp::Ge => *self >= other,
-        }
     }
 
     #[cfg(feature = "std")]
